refactor(tests): refactor models controller tests
